Update README.md
A Blog Website is a web application that allows users to create, publish, edit, and delete blog posts. It serves as an online platform for individuals, communities, and businesses to share news, opinions, and stories. The user interface will be built using HTML, CSS, JavaScript, and Bootstrap to ensure responsiveness and ease of use. The backend will be developed using Laravel (PHP framework) and MySQL for database management. This project aims to provide an organized, accessible, and interactive space for content sharing..